+++ modulo-03-2-springdata/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/model/dto/output/EnderecoOutputDTO.java
@@ -17,7 +17,7 @@ public class EnderecoOutputDTO extends EnderecoInputDTO {
 
     @Nullable
     @Schema(description = "ID da pessoa", example = "2", required = true)
-    private Integer idEndereco;
+    private Long idEndereco;
 
 
 }
