+++ modulo-03-2-springdata/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/model/dto/output/PessoaOutputDTO.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class PessoaOutputDTO extends PessoaInputDTO {
 
     @Schema(description = "ID da pessoa", example = "RESIDENCIAL", required = true)
-    private Integer idPessoa;
+    private Long idPessoa;
 
 
 }
