+++ modulo-03-2-springdata/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/repository/ContatoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContatoRepository extends JpaRepository<Contato, Integer> {
+public interface ContatoRepository extends JpaRepository<Contato, Long> {
 
 //    ContatoR updateByidContato(Integer id, Contato contato);
