+++ modulo-03-2-springdata/aula-01/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/controller/ContatoController.java
@@ -33,7 +33,7 @@ public class ContatoController implements ContatoControllerDoc {
     }
 
     @GetMapping("/pessoa/{idPessoa}")
-    public List<ContatoOutputDTO> contatosByIdPessoa(@Positive @PathVariable("idPessoa") Integer idPessoa) {
+    public List<ContatoOutputDTO> contatosByIdPessoa(@Positive @PathVariable("idPessoa") Long idPessoa) {
         return contatoService.getContatosByIdUsuario(idPessoa);
     }
 
@@ -43,13 +43,13 @@ public class ContatoController implements ContatoControllerDoc {
     }
 
     @PutMapping("/{idContato}")
-    public ResponseEntity<ContatoOutputDTO> update(@Positive @PathVariable("idContato") Integer id,
+    public ResponseEntity<ContatoOutputDTO> update(@Positive @PathVariable("idContato") Long id,
                                                    @Valid @RequestBody ContatoInputDTO contatoAtualizar) throws RegraDeNegocioException, EnumException {
         return ResponseEntity.ok().body(contatoService.update(id, contatoAtualizar));
     }
 
     @DeleteMapping("/{idContato}")
-    public ResponseEntity<Void> delete(@Positive @PathVariable("idContato") Integer id) throws RegraDeNegocioException {
+    public ResponseEntity<Void> delete(@Positive @PathVariable("idContato") Long id) throws RegraDeNegocioException {
         contatoService.delete(id);
         return ResponseEntity.ok().build();
     }
