+++ modulo-03-2-springdata/aula-01/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/model/dto/input/ContatoInputDTO.java
@@ -18,7 +18,7 @@ public class ContatoInputDTO {
 
     @NotNull
     @Schema(description = "Id da pessoa", example = "2", required = true)
-    private Integer idPessoa;
+    private Long idPessoa;
 
     @NotNull
     @NotBlank
