+++ modulo-03-2-springdata/aula-01/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/model/dto/input/EnderecoInputDTO.java
@@ -21,7 +21,7 @@ public class EnderecoInputDTO {
     @Schema(description = "Id de pessoa", example = "RESIDENCIAL", required = true)
     @NotNull
     @Column(name = "id_pessoa")
-    private Integer idPessoa;
+    private Long idPessoa;
 
     @NotNull
     @Schema(description = "Tipo do contato", example = "COMERCIAL", required = true, enumAsRef = true)
