+++ modulo-03-2-springdata/aula-01/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/service/EnderecoService.java
@@ -28,14 +28,14 @@ public class EnderecoService {
         return enderecoRepository.findAll().stream().map(endereco -> objectMapper.convertValue(endereco, EnderecoOutputDTO.class)).toList();
     }
 
-    public List<EnderecoOutputDTO> getEnderecosByIdPessoa(Integer idPessoa) {
+    public List<EnderecoOutputDTO> getEnderecosByIdPessoa(Long idPessoa) {
         return enderecoRepository.findEnderecosByIdPessoa(idPessoa)
                 .stream()
                 .map(endereco -> objectMapper.convertValue(endereco, EnderecoOutputDTO.class))
                 .toList();
     }
 
-    public EnderecoOutputDTO create(Integer idPessoa, EnderecoInputDTO enderecoNovo) throws Exception {
+    public EnderecoOutputDTO create(Long idPessoa, EnderecoInputDTO enderecoNovo) throws Exception {
 
         Pessoa pessoaPorId = objectMapper.convertValue(pessoaService.findById(idPessoa), Pessoa.class);
 
@@ -48,7 +48,7 @@ public class EnderecoService {
         return objectMapper.convertValue(enderecoRepository.save(endereco), EnderecoOutputDTO.class);
     }
 
-    public EnderecoOutputDTO update(Integer idEndereco, EnderecoInputDTO enderecoModificado) throws RegraDeNegocioException, EnumException {
+    public EnderecoOutputDTO update(Long idEndereco, EnderecoInputDTO enderecoModificado) throws RegraDeNegocioException, EnumException {
         this.checarEnum(enderecoModificado.getTipoEndereco());
 
         Endereco enderecoResgatado = enderecoRepository.findById(idEndereco).stream().findFirst().orElseThrow(() -> new RegraDeNegocioException("Não encontrado!"));
@@ -68,12 +68,12 @@ public class EnderecoService {
         return objectMapper.convertValue(enderecoResgatado, EnderecoOutputDTO.class);
     }
 
-    public void delete(Integer idEndereco) throws Exception {
+    public void delete(Long idEndereco) throws Exception {
         Endereco endereco = objectMapper.convertValue(getEnderecoById(idEndereco), Endereco.class);
         enderecoRepository.delete(endereco);
     }
 
-    public EnderecoOutputDTO getEnderecoById(Integer idEndereco) throws RegraDeNegocioException {
+    public EnderecoOutputDTO getEnderecoById(Long idEndereco) throws RegraDeNegocioException {
         Endereco endereco = enderecoRepository.findById(idEndereco).stream().findFirst().orElseThrow(() -> new RegraDeNegocioException("Não encontrado!"));
         return objectMapper.convertValue(endereco, EnderecoOutputDTO.class);
     }
