+++ modulo-03-2-springdata/aula-01/PessoaAPI/src/main/java/br/com/dbc/vemser/pessoaapi/service/PessoaService.java
@@ -39,7 +39,7 @@ public class PessoaService {
                 .toList();
     }
 
-    public PessoaOutputDTO update(Integer id, PessoaInputDTO pessoaAtualizar) throws RegraDeNegocioException, MessagingException {
+    public PessoaOutputDTO update(Long id, PessoaInputDTO pessoaAtualizar) throws RegraDeNegocioException, MessagingException {
         Pessoa pessoaRecuperada = getPessoa(id);
 
         pessoaRecuperada.setCpf(pessoaAtualizar.getCpf());
@@ -54,7 +54,7 @@ public class PessoaService {
         return objectMapper.convertValue(pessoaRecuperada, PessoaOutputDTO.class);
     }
 
-    public void delete(Integer id) throws RegraDeNegocioException, MessagingException {
+    public void delete(Long id) throws RegraDeNegocioException, MessagingException {
         Pessoa pessoaRecuperada = getPessoa(id);
         pessoaRepository.delete(pessoaRecuperada);
 //        emailService.sendTemplateMailDeleteAccount(pessoaRecuperada);
@@ -67,14 +67,14 @@ public class PessoaService {
                 .toList();
     }
 
-    private Pessoa getPessoa(Integer id) throws RegraDeNegocioException {
+    private Pessoa getPessoa(Long id) throws RegraDeNegocioException {
         return pessoaRepository.findById(id)
                 .stream()
                 .findFirst()
                 .orElseThrow(() -> new RegraDeNegocioException("Não encontrou!"));
     }
 
-    public PessoaOutputDTO findById(Integer idBuscado) throws RegraDeNegocioException {
+    public PessoaOutputDTO findById(Long idBuscado) throws RegraDeNegocioException {
         return objectMapper.convertValue(getPessoa(idBuscado), PessoaOutputDTO.class);
     }
 }
